#33 wip split hooks directory
