+++ src/hooks/useTitle.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { useMatches, type UIMatch } from 'react-router';
 import type { RouteHandle } from '../types';
-import { useQueryParams } from './useQueryParams';
+import { useQueryParams } from '../features/currency/hooks/useQueryParams';
 
 export const useTitle = (): string => {
   const matches = useMatches() as UIMatch<unknown, RouteHandle>[];
